+++ CONTRIBUTING.md
@@ -15,18 +15,17 @@ We accept contributions via Pull Requests on [Github](https://github.com/antidot
 
 - **Consider our release cycle** - We try to follow [SemVer v2.0.0](http://semver.org/). Randomly breaking public APIs is not an option.
 
-- **Create feature branches** - Don't ask us to pull from your master branch.
+- **Create feature branches** - Don't ask us to pull from your main branch.
 
 - **One pull request per feature** - If you want to do more than one thing, send multiple pull requests.
 
 - **Send coherent history** - Make sure each individual commit in your pull request is meaningful. If you had to make multiple intermediate commits while developing, please [squash them](http://www.git-scm.com/book/en/v2/Git-Tools-Rewriting-History#Changing-Multiple-Commit-Messages) before submitting.
 
 
-## Running Tests
+## Running Local Pipeline
 
 ``` bash
-$ composer test
+$ composer check-all
 ```
-
 
 **Happy coding**!
