+++ src/Write/Feature.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Pheature\Core\Toggle\Write;
+
+interface Feature
+{
+    public function enable(): void;
+
+    public function disable(): void;
+
+    public function isEnabled(): bool;
+
+    public function id(): string;
+}
